Add helpful hg aliases to show the draft changes in a given tree.
pubbase finds the public change that is the base of the given draft.
gol/gl log all drafts that are in the same tree as the given draft (i.e.
they share the same draft that is a child of their pubbase).
curr is just all of the files modified between the current head and the
pubbase.