Match the rmcup with an smcup to reliably restore the screen after ssh.
Multiple `smcup` calls seem redundant so `rmcup` will go back to the
original before it started. This means that even if the internal `rmcup`
from the remote host is lost (due to hangup or other failure). The outer
`rmcup` will handle it and not leave dangling `cup` state that might be
erroneously brought back after a successful command.