Creates the vim plugin Tedium.
This meant to aid in tedious changes that require opening a lot of
files. For instance adding a new feature to a codebase that requires a
lot of similar configuration changes for each new feature.

It operates on `.ted` files in the ~/.tedium directory.

Each file is in this format:

# Comment
import name
file/path:<ACTION>

Where <ACTION> can be
$ : move to the end of the file.
/regex/ : move to the first occurence of the regex
/regex/} : move to end of the block started by the regex

If the <ACTION> is omitted it is treated as "move to the start of the
file."

import loads another ".ted" file.

Usage:
:Tedi name
# Loads the file ~/.tedium/name.ted

:Ted
# Goes to the next entry in ~/.tedium/name.ted