Add mergediffs vim plugin.
Makes it easy when doing a merge (resolve) in git and hg to view diff of
only a particular buffer against the base.

DiffOther - shows diff of base to the OTHER or REMOTE file.
DiffLocal - shows diff of base to the LOCAL file.
DiffAll - diff all files (this is the default)

NOTE: In rebase LOCAL and REMOTE are swapped because rebase operates by
checking out the other side and then replaying the local side on it.