Add a workaround to the netrw "[No Name]" bug.
This issue happens only when 'hidden' is true. It seems that netrw will
create a new buffer to load its browser. However, when opening a
directory vim will create an empty buffer with that directory's name.

If 'nohidden' then this buffer is deleted as soon as the netrw buffer is
created, this allows netrw to rename its buffer to the directory's name
and everything works fine.

If 'hidden' then the pre-existing empty buffer hangs around and
interferes with the attempt to rename the netrw buffer. This fix just
deletes the empty buffer and renames the netrw buffer.