Wrap the remote call to screen in scr with a zsh login shell.
This allows the zsh configured environment to be set before the screen
command is selected. If not for this, then the ssh environment would be
used, which doesn't usually include the path to homebrew so the builtin
macOS version of screen is used. For some reason macOS still ships with
a version of screen for 2006!

My previous solution to this was to use a custom ssh environment but
this requires changing sshd_config and maintaining a specific ssh
environment in addition to system environment updates. This way, I can
just configure homebrew for the system and it will also be accessible to
the scr command.