Add wrap_format to make \= formatting work better. Also switch to ktfmt.
wrap_format runs the wrapped autoformatter and intelligently uses its
output. If the formatter succeeds, the output is used. If the formatter
fails, then the file is left unmodified and the errors from the
formatter are loaded into vim's qf buffer so that they get immediately
highlighted.

This is necessary because many formatters will output the error when
they fail to parse but not the original file. This way we keep the file
and just jump to the spot where the error was.

This assumes that stderr is where errors are dumped and that a non-zero
exit accompanies them.

ktfmt seems to be better than ktlint. ktlint seems to return 0 on parse
failure if the file comes in via stdin. Also, it doesn't have a way to
override the name that is used for stdin so the error file in qf doesn't
seem to help much. ktfmt doesn't have any of these issues.