Add update_ssh_agent a tool for SSH_AUTHS_SOCK work in a remote screen
This is a best-effort attempt to set a valid agent SSH_AUTH_SOCK if the
current one is empty or invalid. This has limitations. It won't work if
the screen session uses a valid socket that just happens to not include
necessary credentials. In this case, you must manually delete the socket
and then run this. If this happens, you may want to restore the old
socket so that subsequent local uses of the screen session will still
work.